Move process_gcm_data to gcm_climate (#606)
* Move process_gcm_data to climate_prepro

* Rename to gcm_climate
