+++ oggm/tests/test_prepro.py
@@ -19,7 +19,7 @@ import xarray as xr
 import rasterio
 
 # Local imports
-from oggm.core import (gis, inversion, climate_prepro, climate, centerlines,
+from oggm.core import (gis, inversion, gcm_climate, climate, centerlines,
                        flowline, massbalance)
 import oggm.cfg as cfg
 from oggm import utils
@@ -2349,7 +2349,7 @@ class TestGCMClimate(unittest.TestCase):
         cfg.PATHS['cesm_precc_file'] = f
         f = get_demo_file('cesm.PRECL.160001-200512.selection.nc')
         cfg.PATHS['cesm_precl_file'] = f
-        climate_prepro.process_cesm_data(gdir)
+        gcm_climate.process_cesm_data(gdir)
 
         fh = gdir.get_filepath('climate_monthly')
         fcesm = gdir.get_filepath('gcm_data')
@@ -2405,7 +2405,7 @@ class TestGCMClimate(unittest.TestCase):
         cfg.PATHS['cesm_precc_file'] = f
         f = get_demo_file('cesm.PRECL.160001-200512.selection.nc')
         cfg.PATHS['cesm_precl_file'] = f
-        climate_prepro.process_cesm_data(gdir, filesuffix=filesuffix)
+        gcm_climate.process_cesm_data(gdir, filesuffix=filesuffix)
         utils.compile_climate_input([gdir], filename=filename,
                                     filesuffix=filesuffix)
